Added IDBAdapterFactory and fixed tests
Adding IDBAdapterFactory interface so that specific implementations can
control the types of SharedDB and DedicatedDB adapters are created.
This is also useful for testing as we can create mock factories that
create adapters that logically simulate DB interactions. This was also
needed as some syntax used in MySQL and Postgres is not supported by
sqlite. Therefore, tests would fail once things like CREATE DATABASE
were executed on sqlite but would work in production in production.