modernized pipeline.
* updated `aggregate` to `in_parallel`

* removed yaml anchors, not working in concourse 5.x.

* smoke tests now use 18f/cg-labratory/databases/aws-rds db verification
app. it migrates the smoke tests from self-compiled shell scripts to a
golang app, which includes the supported database drivers. the app
replicates the shell scripts, using the same SQL statements, so they are
functionally the same. by updating this, it removes the dependencies on
self-compiled shell scripts, and the aws-rds app can be manually pushed,
if needed.

* removed the self-compiled shell scripts for various db clients.

* smoke tests now push a non-started version of aws-rds, set environment
variables to reflect the database configs, bind the service, then
re-push to stage the app.

* pipeline now runs on 18fgsa/sql-clients container image, which is used
for troubleshooting issues with rds dbs, it has all the requisite
clients needed.

* updated golang image from 1.8-wheezy to latest.

Signed-off-by: Mike Lloyd <mike.lloyd@gsa.gov>